Use `Long` for anything counting reads

+++ src/main/scala/org/broadinstitute/gpp/poolq3/hist/Histogram.scala
@@ -8,12 +8,12 @@ package org.broadinstitute.gpp.poolq3.hist
 trait ReadOnlyHistogram[A]:
 
   /** Returns the number of occurrences of key `k` */
-  def count(k: A): Int
+  def count(k: A): Long
 
   /** Returns the keys tracked in this histogram */
   def keys: Set[A]
 
-  def toMap: scala.collection.immutable.Map[A, Int] = keys.map(k => k -> count(k)).toMap
+  def toMap: scala.collection.immutable.Map[A, Long] = keys.map(k => k -> count(k)).toMap
 
 end ReadOnlyHistogram
 
@@ -21,4 +21,4 @@ end ReadOnlyHistogram
 trait Histogram[A] extends ReadOnlyHistogram[A]:
 
   /** Increment the occurrences of key `k` */
-  def increment(k: A): Int
+  def increment(k: A): Long
